Issue #3891: Split and Organize Checkstyle inputs by Test for ArrayTrailingComma.
