+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/ArrayTrailingCommaCheckTest.java
@@ -35,7 +35,7 @@ public class ArrayTrailingCommaCheckTest
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "coding" + File.separator + filename);
+                + "coding" + File.separator + "arraytrailingcomma" + File.separator + filename);
     }
 
     @Test
